style : 스토리북 - 프리텐다드 폰트 적용, 기본 레이아웃 스타일 적용
